NEW Post specific permission setting
Allow users to set the permissions they want on a per-post basis when posting.

Makes use of a "Logged in users" option, that is configured on a per-site or
per-system basis, letting an admin define what the 'logged in' users groups
are.

Additionally, allows for selecting a specific user or group to send the post
to.