NEW 'Typed' posts can have different ages
If a post is created with a particular "PostType", it can be removed from
feeds by specifying the `typeAge` variable. For example a 'notice' type post
could be displayed for only an hour before it drops out of display, whereas
general posts will continue displaying indefinitely.

YML config:

```yml
Injector:
  MicroBlogService:
    properties:
      typeAge:
        'notice-post': 1800
```